moved OAuth2TokenIntrospectionClaimsHttpMessageConverter and removed HttpMessageConverters
